git commit -m "Minor bugs
There was an issue with the end condition where it would continuously rotate even after detection reads >30 but I couldnt replicate it and it shouldn't be a problem in the maze

Changed center range to give more tolerance for the angle it can approach the wall. Needs more testing to see if necessary. Prefer the range to be as tight as possible."