Final testing updates:
This code was intented to be used for the demo. It managed to complete
the maze autonomously 3 times, however it took too long and had too many
collisions, resulting in too many marks lost. A cheese version was made
for the demo. This version was going to be used for the autonomous bonus
test but it failed BECAUSE THE BATTERY CHARGE WAS DIFFERENT FROM TESTING
CAUSING IT TO OVERSHOOT THE FIRST CELL. So stupid. Honestly could have
used the second version and done just fine due to how the maze was
shaped because there was a piece of code I added there that I forgot to
add to prevent issues when hitting walls. Will make changes for final
upload in the future."