Cheese version of maze solving algorithm.
This version was used in the demo. It was made to overshoot the cell by a bit in order to skip the paths leading to the dead end. The algorithm was also updated to check forward everytime regardless in order to recenter itself in the cell. This should always have been the case but was made necessary upon realizing that it will detect the poll holding up the wall when pressed right up against the front wall resulting in it thinking there is a wall to the left even if there isnt. Readjusting fixes that issue.

The robot managed to clear the maze in 33 seconds with 2 collisions. One of the collisions could have been prevented by overshooting the entry cell and having the adjustment check pull in to the second cell perfectly instead of having to crash into it.